+++ PureInvoke/Functions/Invoke-AdvapiLookupAccountName.ps1
@@ -1,12 +1,12 @@
 
-function Invoke-PLookupAccountName
+function Invoke-AdvapiLookupAccountName
 {
     <#
     .SYNOPSIS
     Calls the Advanced Windows 32 Base API (advapi32.dll) `LookupAccountName` function.
 
     .DESCRIPTION
-    The `Invoke-PLookupAccountName` function calls the advapi32.dll API's `LookupAccountName` function, which looks up
+    The `Invoke-AdvapiLookupAccountName` function calls the advapi32.dll API's `LookupAccountName` function, which looks up
     an account name and returns its domain, SID, and use. Pass the account name to the `AccountName` parameter and the
     system name to the `SystemName` parameter, which are passed to `LookupAccountName` as the `lpAccountName` and
     `lpSystemName` arguments, respectively. The function returns an object with properties for each of the
@@ -16,7 +16,7 @@ function Invoke-PLookupAccountName
     https://learn.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-lookupaccountnamea
 
     .EXAMPLE
-    Invoke-PLookupAccountName -AccountName ([Environment]::UserName)
+    Invoke-AdvapiLookupAccountName -AccountName ([Environment]::UserName)
 
     Demonstrates how to call this function by passing a username to the `AccountName` parameter.
     #>
@@ -36,7 +36,7 @@ function Invoke-PLookupAccountName
     $result = [pscustomobject]@{
         ReferencedDomainName = '';
         Sid = [byte[]]::New(0);
-        Use = [PInvoke.AdvApi32+SidNameUse]::Unknown
+        Use = [PureInvoke.AdvApi32+SidNameUse]::Unknown
     }
 
     [byte[]] $sid = [byte[]]::New(0);
@@ -46,34 +46,34 @@ function Invoke-PLookupAccountName
     [Text.StringBuilder] $domainName = [Text.StringBuilder]::New()
     # cch = count of chars
     [UInt32] $cchDomainName = $domainName.Capacity;
-    [PInvoke.AdvApi32+SidNameUse] $sidNameUse = [PInvoke.AdvApi32+SidNameUse]::Unknown;
+    [PureInvoke.AdvApi32+SidNameUse] $sidNameUse = [PureInvoke.AdvApi32+SidNameUse]::Unknown;
 
-    $err = [PInvoke.WinError]::Ok
-    if ([PInvoke.AdvApi32]::LookupAccountName($SystemName, $AccountName,
+    $err = [PureInvoke.WinError]::Ok
+    if ([PureInvoke.AdvApi32]::LookupAccountName($SystemName, $AccountName,
                                               $sid, [ref] $cbSid,
                                               $domainName, [ref] $cchDomainName,
                                               [ref]$sidNameUse))
     {
-        Write-PWin32Error
+        Write-Win32Error
         return
     }
 
     $err = [Marshal]::GetLastWin32Error();
-    if ($err -eq [PInvoke.WinError]::InsufficientBuffer -or $err -eq [PInvoke.WinError]::InvalidFlags)
+    if ($err -eq [PureInvoke.WinError]::InsufficientBuffer -or $err -eq [PureInvoke.WinError]::InvalidFlags)
     {
         $sid = [byte[]]::New($cbSid);
         [void]$domainName.EnsureCapacity([int]$cchDomainName);
-        if (-not [PInvoke.AdvApi32]::LookupAccountName($SystemName, $AccountName, $sid, [ref] $cbSid,
+        if (-not [PureInvoke.AdvApi32]::LookupAccountName($SystemName, $AccountName, $sid, [ref] $cbSid,
                                                        $domainName, [ref] $cchDomainName,
                                                        [ref] $sidNameUse))
         {
-            Write-PWin32Error
+            Write-Win32Error
             return
         }
     }
     else
     {
-        Write-PWin32Error
+        Write-Win32Error
         return
     }