+++ PureInvoke/Functions/Invoke-AdvapiLookupAccountSid.ps1
@@ -1,12 +1,12 @@
 
-function Invoke-PLookupAccountSid
+function Invoke-AdvapiLookupAccountSid
 {
     <#
     .SYNOPSIS
     Calls the Advanced Windows 32 Base API (advapi32.dll) `LookupAccountSid` function.
 
     .DESCRIPTION
-    The `Invoke-PLookupAccountSid` function calls the advapi32.dll API's `LookupAccountSid` function, which looks up a
+    The `Invoke-AdvapiLookupAccountSid` function calls the advapi32.dll API's `LookupAccountSid` function, which looks up a
     SID and returns its account name, domain name, and use. Pass the SID as a byte array to the `Sid` parameter and the
     system name to the `SystemName` parameter, which are passed to `LookupAccountSid` as the `Sid` and `lpSystemName`
     arguments, respectively. The function returns an object with properties for each of the `LookupAccountSid`
@@ -16,7 +16,7 @@ function Invoke-PLookupAccountSid
     https://learn.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-lookupaccountsida
 
     .EXAMPLE
-    Invoke-PLookupAccountSid -Sid $sid
+    Invoke-AdvapiLookupAccountSid -Sid $sid
 
     Demonstrates how to call this function by passing a sid to the `Sid` parameter.
     #>
@@ -34,7 +34,7 @@ function Invoke-PLookupAccountSid
     $result = [pscustomobject]@{
         Name = '';
         ReferencedDomainName = ''
-        Use = [PInvoke.AdvApi32+SidNameUse]::Unknown
+        Use = [PureInvoke.AdvApi32+SidNameUse]::Unknown
     }
 
     [Text.StringBuilder] $name = [Text.StringBuilder]::New()
@@ -44,20 +44,20 @@ function Invoke-PLookupAccountSid
     [Text.StringBuilder] $domainName = [Text.StringBuilder]::New()
     [UInt32] $cchDomainName = $domainName.Capacity;
 
-    [PInvoke.AdvApi32+SidNameUse] $sidNameUse = [PInvoke.AdvApi32+SidNameUse]::Unknown;
+    [PureInvoke.AdvApi32+SidNameUse] $sidNameUse = [PureInvoke.AdvApi32+SidNameUse]::Unknown;
 
-    $err = [PInvoke.WinError]::Ok
+    $err = [PureInvoke.WinError]::Ok
 
-    if (-not ([PInvoke.AdvApi32]::LookupAccountSid($SystemName, $sid, $name, [ref] $cchName,
+    if (-not ([PureInvoke.AdvApi32]::LookupAccountSid($SystemName, $sid, $name, [ref] $cchName,
                                                    $domainName, [ref] $cchDomainName, [ref] $sidNameUse)))
     {
         $err = [Marshal]::GetLastWin32Error();
-        if ($err -eq [PInvoke.WinError]::InsufficientBuffer)
+        if ($err -eq [PureInvoke.WinError]::InsufficientBuffer)
         {
             [void]$name.EnsureCapacity([int]$cchName);
             [void]$domainName.EnsureCapacity([int]$cchName);
             $err = 0
-            if (-not [PInvoke.AdvApi32]::LookupAccountSid($SystemName, $sid,  $name, [ref] $cchName,
+            if (-not [PureInvoke.AdvApi32]::LookupAccountSid($SystemName, $sid,  $name, [ref] $cchName,
                                                         $domainName, $cchDomainName, [ref] $sidNameUse))
             {
                 $err = [Marshal]::GetLastWin32Error()
@@ -67,7 +67,7 @@ function Invoke-PLookupAccountSid
 
     if ($err)
     {
-        Write-PWin32Error
+        Write-Win32Error
         return
     }